perf(@ngtools/webpack): reduce rebuild performance by typechecking more (#4258)
We got the TypeScript performance down a lot by not type checking every files on rebuild, but this has an issue where typings can be wrong in files that depends on us. This PR alleviate the problem by type checking all files that depends on the changed file.

Note that even if this PR reduces performance, we're still ~40% faster
than Beta.26.