fix(@angular/cli): remove broken script after eject
`prepree2e` doesn't run because it's a lifecycle script for a lifecycle script, but even if it did it would not work because there is nothing coordinating the serve and e2e processes.

Users must manually initialize the server prior to running e2e tests after ejecting.

Related to https://github.com/angular/angular-cli/issues/6171