fix(@angular/cli): Make generated inline template conform to test case
The generated app component spec checks for the string "Welcome to {{title}}!!", which is only output by the generated `app.component.htm`l file, not the generated inline template in `app.component.ts` when the `--inline-template` option is passed to the `ng new` command. This causes a test failure when generating a new app with `--inline-template`.