fix(@ngtools/webpack): fix bootstrapping code
There is no link between a new identifier and its import (and apparently TS does
not export any API to make one in the type checker). The best way to import a
symbol reliably is to use a namespace import (import * as x from "abc").

The new bootstrapping code uses just that.